moved test to test to follow file structure for project
